get access to public methods of the hangups Client
the client methods can be accessed via ```bot._client.<method>``` and now also via ```bot.<method>```

If a method with the same name is implemented in the ```HangupsBot``` class, the method will be used instead.

e.x.:
calling ```bot._client.upload_image(<args>)``` is an access to a private member of ```bot```, however the method ```image_upload``` is a common method and an access should not violate any permission. It can be accessed now via ```bot.image_upload```.