[plugins.telepot] keep up with aiohttp changes
use das7pads v12.6 telepot version:
see https://github.com/das7pad/telepot/releases/tag/v12.6.1

This release is a bugfix release for the upstream release v12.6.

It covers the changes needed to keep up with the promotion of aiohttp.ClientSession.close to a coroutine.
The telepot.aio.Bot class has a new coroutine close to cleanup the underlaying connection pool. It is a shotcut to telepot.aio.api.close_pools. The close method may be extended with further cleanup.